fix: wrap total_return_pct in float() for consistent numeric type
The scoped re-review of the final-review fix wave found this was the
one leftover spot from the C4 rewrite where numpy dtypes (from real
close prices) could leak into the metrics dict instead of plain
Python floats, unlike realized_pnl/unrealized_pnl/total_pnl which
were already wrapped.

+++ strategy_engine/backtest/metrics.py
@@ -36,5 +36,5 @@ def compute_metrics(closed_trades: list, open_positions=(), current_prices=None)
         "unrealized_pnl": float(unrealized_pnl),
         "total_pnl": float(total_pnl),
         "win_rate": (wins / len(closed_trades)) if closed_trades else 0.0,
-        "total_return_pct": (total_pnl / total_invested) if total_invested else 0.0,
+        "total_return_pct": float(total_pnl / total_invested) if total_invested else 0.0,
     }
